requirements.txt
Flask==2.3.2
pymongo==4.6.0
python-dotenv==1.0.0
bcrypt==4.0.1